requirements.txt
# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Core trading bot dependencies
smartapi-python==1.4.7
logzero>=1.7.0
pyotp>=2.8.0
requests>=2.31.0
websocket-client>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Machine Learning dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
# tensorflow>=2.13.0  # Optional - comment out if causing issues
xgboost>=1.7.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Technical analysis (fixed version)
pandas-ta==0.4.71b0

# Scientific computing
scipy>=1.11.0

# Date/time handling
python-dateutil>=2.8.2